requirements.txt
Django==1.8.6
django-haystack==2.4.0
django-taggit==0.17.1
Markdown==2.6.2
pysolr==3.3.2
pytz==2016.10
requests==2.13.0
